requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0

# Data generation
faker>=18.0.0

# Machine Learning / GNN
torch>=2.0.0
torch-geometric>=2.3.0
scikit-learn>=1.3.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pgvector>=0.2.0

# NLP and embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
tiktoken>=0.4.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0

# Development/Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
